+++ packages/ai/src/ButtonState.ts
@@ -11,8 +11,8 @@ import property from "@ui5/webcomponents-base/dist/decorators/property.js";
  *
  * ### Usage
  *
- * `ui5-si-button-state` is an abstract element, representing a state of `ui5-ai-button`. It is meant to be used in the `states` slot
- * of `ui5-ai-button` and should be used as standalone component.
+ * `ui5-ai-button-state` is an abstract element, representing a state of `ui5-ai-button`. It is meant to be used in the `default` slot
+ * of `ui5-ai-button` and should not be used as standalone component.
  *
  * ### ES6 Module Import
  *
@@ -22,7 +22,7 @@ import property from "@ui5/webcomponents-base/dist/decorators/property.js";
  * @abstract
  * @since 2.0.0
  * @public
- * @experimental The Button and ButtonState web components are availabe since 2.0 under an experimental flag and their API and behaviour are subject to change.
+ * @experimental The Button and ButtonState web components are available since 2.0 under an experimental flag and their API and behaviour are subject to change.
  */
 @customElement("ui5-ai-button-state")
 class ButtonState extends UI5Element {