+++ packages/main/src/Select.ts
@@ -168,9 +168,7 @@ type SelectLiveChangeEventDetail = {
  * Fired after the component's dropdown menu opens.
  * @public
  */
-@event("open", {
-	bubbles: true,
-})
+@event("open")
 
 /**
  * Fired after the component's dropdown menu closes.
@@ -365,7 +363,11 @@ class Select extends UI5Element implements IFormInputElement {
 		const selectedOption = this.selectedOption;
 
 		if (selectedOption) {
-			return selectedOption.hasAttribute("value") ? selectedOption.value! : selectedOption.textContent;
+			if ("value" in selectedOption && selectedOption.value) {
+				return selectedOption.value;
+			}
+
+			return selectedOption.hasAttribute("value") ? selectedOption.getAttribute("value") : selectedOption.textContent;
 		}
 
 		return "";