+++ packages/main/src/TableHeaderCell.ts
@@ -1,8 +1,12 @@
-import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import property from "@ui5/webcomponents-base/dist/decorators/property.js";
+import { customElement, property, slot } from "@ui5/webcomponents-base/dist/decorators.js";
 import TableCellBase from "./TableCellBase.js";
 import TableHeaderCellTemplate from "./generated/templates/TableHeaderCellTemplate.lit.js";
 import TableHeaderCellStyles from "./generated/themes/TableHeaderCell.css.js";
+import Icon from "./Icon.js";
+import SortOrder from "@ui5/webcomponents-base/dist/types/SortOrder.js";
+import type TableHeaderCellActionBase from "./TableHeaderCellActionBase.js";
+import "@ui5/webcomponents-icons/dist/sort-ascending.js";
+import "@ui5/webcomponents-icons/dist/sort-descending.js";
 
 /**
  * @class
@@ -28,6 +32,7 @@ import TableHeaderCellStyles from "./generated/themes/TableHeaderCell.css.js";
 	tag: "ui5-table-header-cell",
 	styles: [TableCellBase.styles, TableHeaderCellStyles],
 	template: TableHeaderCellTemplate,
+	dependencies: [Icon],
 })
 class TableHeaderCell extends TableCellBase {
 	/**
@@ -85,6 +90,40 @@ class TableHeaderCell extends TableCellBase {
 	@property()
 	popinText?: string;
 
+	/**
+	 * Defines the sort indicator of the column.
+	 *
+	 * @default "None"
+	 * @since 2.8.0
+	 * @public
+	 */
+	@property()
+	sortIndicator: `${SortOrder}` = "None";
+
+	/**
+	 * Defines if the column is hidden in the popin.
+	 *
+	 * **Note:** Please be aware that hiding the column in the popin might lead to accessibility issues as
+	 * users might not be able to access the content of the column on small screens.
+	 *
+	 * @default false
+	 * @since 2.8.0
+	 * @public
+	 */
+	@property({ type: Boolean })
+	popinHidden: boolean = false;
+
+	/**
+	 * Defines the action of the column.
+	 *
+	 * **Note:** While multiple actions are technically possible, this is not supported.
+	 *
+	 * @public
+	 * @since 2.8.0
+	 */
+	@slot()
+	action!: Array<TableHeaderCellActionBase>;
+
 	@property({ type: Boolean, noAttribute: true })
 	_popin = false;
 
@@ -103,6 +142,17 @@ class TableHeaderCell extends TableCellBase {
 		if (this._individualSlot) {
 			// overwrite setting of TableCellBase so that the TableHeaderCell always uses the slot variable
 			this.style.justifyContent = `var(--horizontal-align-${this._individualSlot})`;
+		}
+		if (this.sortIndicator !== SortOrder.None) {
+			this.setAttribute("aria-sort", this.sortIndicator.toLowerCase());
+		} else if (this.hasAttribute("aria-sort")) {
+			this.removeAttribute("aria-sort");
+		}
+	}
+
+	get _sortIcon() {
+		if (this.sortIndicator !== SortOrder.None) {
+			return `sort-${this.sortIndicator.toLowerCase()}`;
 		}
 	}
 }