+++ packages/main/src/TableRow.ts
@@ -48,7 +48,7 @@ class TableRow extends TableRowBase {
 		"default": true,
 		individualSlots: true,
 		invalidateOnChildChange: {
-			properties: ["_popin"],
+			properties: ["_popin", "_popinHidden"],
 			slots: false,
 		},
 	})
@@ -71,14 +71,16 @@ class TableRow extends TableRowBase {
 	/**
 	 * Unique identifier of the row.
 	 *
-	 * @default ""
+	 * **Note:** For selection features to work properly, this property is mandatory, and its value must not contain spaces.
+	 *
+	 * @default undefined
 	 * @public
 	 */
 	@property()
-	rowKey = "";
+	rowKey?: string;
 
 	/**
-	 * Defines the position of the row respect to the total number of rows within the table when the `ui5-table-virtualizer` feature is used.
+	 * Defines the position of the row related to the total number of rows within the table when the `ui5-table-virtualizer` feature is used.
 	 *
      * @default -1
 	 * @since 2.5.0